refactor(enrichers): rename company_enricher to company_fuzzy_matcher
Every module in sbir_etl/enrichers/ is an "enricher", so naming one
of them `company_enricher.py` conveys nothing about what it does.
Worse, it sits next to `company_enrichment.py` (federal-API lookups)
which is a completely different responsibility — the two names were
indistinguishable and the docs/architecture file even mislabeled
this one as "SAM.gov company enrichment", which it isn't.

Rename to `company_fuzzy_matcher.py` to make the responsibility
explicit: this module does pure local rapidfuzz matching against a
company master list, with no external API calls.

Updates:
- git mv source and test files (history preserved)
- 3 production importers (examples/, scripts/, utils/company_canonicalizer)
- text_normalization.py docstrings (4 references)
- config/base.yaml comment
- docs/architecture/detailed-overview.md (also corrected the wrong
  "SAM.gov company enrichment" comment)
- docs/steering/structure.md snake_case example
- docs/enrichment/enhanced-matching.md import example, plus a stale
  test file reference (test_company_enricher_enhanced.py never existed)
- .secrets.baseline path

Archived docs/status reports under docs/archive/ and tests/archive/
are intentionally left alone — they describe historical state.

No behavior change. All 607 enricher unit tests still pass.
