+++ .secrets.baseline
@@ -154,10 +154,10 @@
         "line_number": 269
       }
     ],
-    "tests/unit/enrichers/test_company_enricher.py": [
+    "tests/unit/enrichers/test_company_fuzzy_matcher.py": [
       {
         "type": "Hex High Entropy String",
-        "filename": "tests/unit/enrichers/test_company_enricher.py",
+        "filename": "tests/unit/enrichers/test_company_fuzzy_matcher.py",
         "hashed_secret": "0091061a3babbe6f11d48aa0142e22341b3ea446",
         "is_verified": false,
         "line_number": 25
