+++ sbir_etl/utils/text_normalization.py
@@ -1,7 +1,7 @@
 """Text normalization utilities for consistent name matching across enrichers.
 
 This module provides shared text normalization functions to eliminate duplicate
-normalization logic across company_enricher, usaspending_enricher, and other modules.
+normalization logic across company_fuzzy_matcher, usaspending_enricher, and other modules.
 
 Key Features:
 - Unified name normalization with configurable suffix handling
@@ -26,7 +26,7 @@ def normalize_name(
     """Normalize a company or recipient name for fuzzy matching.
 
     This function provides unified normalization logic used across multiple enrichers:
-    - company_enricher: Uses normalize_suffixes=True (keeps standardized suffixes)
+    - company_fuzzy_matcher: Uses normalize_suffixes=True (keeps standardized suffixes)
     - usaspending_enricher: Uses remove_suffixes=True (strips all suffixes)
 
     Args:
@@ -70,7 +70,7 @@ def normalize_name(
             s,
         )
     else:
-        # Normalize suffixes to standard forms (company_enricher behavior)
+        # Normalize suffixes to standard forms (company_fuzzy_matcher behavior)
         s = re.sub(r"\b(incorporated|incorporation)\b", "inc", s)
         s = re.sub(r"\b(company|co)\b", "company", s)
         s = re.sub(r"\b(limited|ltd)\b", "ltd", s)
@@ -102,7 +102,7 @@ def normalize_company_name(name: str | None) -> str:
     """Normalize a company name (keeps standardized suffixes).
 
     This is a backward-compatible wrapper for normalize_name() with
-    remove_suffixes=False. Used by company_enricher.
+    remove_suffixes=False. Used by company_fuzzy_matcher.
 
     Args:
         name: Company name to normalize
